Several changes (#18)
Show help if no argument is provided

In case someone calls `xmppc` assume that he doesn't know how to use it.
Which means he might even not be aware of `-h`.
So let's print the help in case no argument is provided instead of
connecting, then discoverin that no mode is selected and disconnecting
again.

Fix segfault when not roster subcommand is provided

_roster_parse_command() segfaults since it assumes there is an argv[0].
Let's test the number of arguments.

Since we are already connected xmpp_connect_client() will wait for a
disconnect. So I introduce a teardown function that is called in both
the error case and the succesful case.

Now _roster_send_query() doesn't disconnect anymore.

Set to development by default

Only set to release when a new release is tagged.
Like this not every developer needs to change the code.
And it's git aka development anyways.

Don't mix german and english

Co-authored-by: Michael Vetter <jubalh@iodoru.org>
Reviewed-on: https://codeberg.org/Anoxinon_e.V./xmppc/pulls/18